move BuildConfig function into viper file + simplifying destination blockchain overwrite
